refactor text directory
