+++ src/text/process-commands/channel/channel-validation.ts
@@ -1,9 +1,9 @@
 import { l, err } from '@/logging'
 import type { ProcessingOptions } from '@/types'
 
-export function validateRSSOptions(options: ProcessingOptions): void {
-  const p = '[text/process-commands/rss/validation]'
-  l.dim(`${p} Starting RSS options validation`)
+export function validateChannelOptions(options: ProcessingOptions): void {
+  const p = '[text/process-commands/channel/channel-validation]'
+  l.dim(`${p} Starting channel options validation`)
   
   if (options.last !== undefined) {
     if (!Number.isInteger(options.last) || options.last < 1) {
@@ -16,12 +16,12 @@ export function validateRSSOptions(options: ProcessingOptions): void {
     }
     l.dim(`${p} Last option validated: ${options.last}`)
   }
-  
+
   if (options.order !== undefined && !['newest', 'oldest'].includes(options.order)) {
     err("Error: The --order option must be either 'newest' or 'oldest'.")
     process.exit(1)
   }
-  
+
   if (options.days !== undefined) {
     if (!Number.isInteger(options.days) || options.days < 1) {
       err('Error: The --days option must be a positive integer.')
@@ -37,7 +37,7 @@ export function validateRSSOptions(options: ProcessingOptions): void {
     }
     l.dim(`${p} Days option validated: ${options.days}`)
   }
-  
+
   if (options.date && options.date.length > 0) {
     const dateRegex = /^\d{4}-\d{2}-\d{2}$/
     for (const d of options.date) {
@@ -55,6 +55,14 @@ export function validateRSSOptions(options: ProcessingOptions): void {
     }
     l.dim(`${p} Date filtering enabled for ${options.date.length} dates: ${options.date.join(', ')}`)
   }
+
+  if (options.last) {
+    l.dim(`\nProcessing the last ${options.last} videos`)
+  } else if (options.days) {
+    l.dim(`\nProcessing videos from the last ${options.days} days`)
+  } else if (options.date && options.date.length > 0) {
+    l.dim(`\nProcessing videos from specific dates: ${options.date.join(', ')}`)
+  }
   
-  l.dim(`${p} RSS validation completed successfully`)
+  l.dim(`${p} Channel validation completed successfully`)
 }
