+++ src/text/process-commands/rss/index.ts
@@ -1,5 +1,4 @@
-import { validateRSSOptions } from './validation.ts'
-import { getLLMService, getTranscriptService } from './services.ts'
+import { validateRSSOptions, getLLMService, getTranscriptService } from './rss-validation.ts'
 import { processRSSFeeds } from './processor.ts'
 import { handleWorkflow } from './workflows.ts'
 import { l, err, logInitialFunctionCall } from '@/logging'
